pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ollama-agent-kit"
version = "0.1.0"
description = "Teaching-oriented Ollama agent with tool calling and CLI chat"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27,<1",
    "python-dotenv>=1.0,<2",
    "rich>=14,<15",
    "typer>=0.12,<1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "ruff>=0.11,<0.12",
]

[project.scripts]
ollama-agent = "ollama_agent_kit.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
